% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_groups_cts.R
\name{additional_circle_centers}
\alias{additional_circle_centers}
\title{Finds additional centers for 2d circles}
\usage{
additional_circle_centers(samples, centers, radius, gsize, TOL = 1e-05)
}
\arguments{
\item{samples}{1 x k x d array of post. samples}

\item{centers}{1 x k x d array of centers}

\item{radius}{radius of circles}

\item{gsize}{grid size}
}
\description{
Finds additional centers for 2d circles
}
\keyword{internal}
