% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blip.R
\name{binarize_selections}
\alias{binarize_selections}
\title{Have to decide if this is internal or exported}
\usage{
binarize_selections(
  cand_groups,
  q,
  v_opt,
  error,
  deterministic,
  tol = 0.001,
  verbose = F
)
}
\description{
Have to decide if this is internal or exported
}
