% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blip.R
\name{BLiP_cts}
\alias{BLiP_cts}
\title{BLiP when the set of locations is continuous, e.g., when
working with image data.}
\usage{
BLiP_cts(
  locs,
  grid_sizes = DEFAULT_GRID_SIZES,
  weight_fn = inverse_radius,
  extra_centers = NULL,
  max_pep = 0.25,
  shape = "circle",
  min_blip_size = 5000,
  verbose = T,
  ...
)
}
\arguments{
\item{locs}{A (N, num_disc, d)-dimensional array. Here, N is the
number of samples from the posterior, d is the number
of dimensions of the space, and each point corresponds
to a signal in a particular posterior sample.}

\item{grid_sizes}{List of grid sizes to split up the locations.
The grid size is inversely proportional to the distance between
lattice points.}

\item{weight_fn}{Weight function which maps candidate groups
to weights. Defaults to inverse radius.}

\item{extra_centers}{An (ncenters, d)-dimensional matrix. At
each resolution, candidate groups will be computed with centers
at this location.}

\item{max_pep}{The maximum allowable PEP for output candidate groups.
Defaults to 0.25.}

\item{shape}{One of ('circle', 'square').}

\item{min_blip_size}{Combines connected components so all subproblems
are at least this size.}

\item{verbose}{If T, gives occasional status updates. Defaults to T.}

\item{...}{Other arguments to the underlying BliP call,
such as the error rate or nominal level.}
}
\description{
BLiP when the set of locations is continuous, e.g., when
working with image data.
}
