% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_groups_cts.R
\name{broadcast_diffs}
\alias{broadcast_diffs}
\title{#' @keywords internal}
\usage{
broadcast_diffs(M1, M2)
}
\arguments{
\item{M1}{a k1 x d matrix}

\item{M2}{a k2 x d matrix}
}
\value{
a k1 x k2 x d array of differences
}
\description{
#' @keywords internal
}
\keyword{internal}
