% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_groups_cts.R
\name{centerrad2key}
\alias{centerrad2key}
\title{Maps (center, rad) pairs to string keys.}
\usage{
centerrad2key(centers, rad, dec = NULL)
}
\arguments{
\item{centers}{1 x k x d array of centers}

\item{rad}{radius}
}
\description{
Maps (center, rad) pairs to string keys.
}
\keyword{internal}
