% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_groups.R
\name{dist_matrix_to_groups}
\alias{dist_matrix_to_groups}
\title{Creates groups based on dist_matrix using hierarchical clustering}
\usage{
dist_matrix_to_groups(dist_matrix)
}
\description{
Creates groups based on dist_matrix using hierarchical clustering
}
