% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecc.R
\name{edge_clique_cover}
\alias{edge_clique_cover}
\title{Finds an edge clique cover of a graph G.}
\usage{
edge_clique_cover(G)
}
\arguments{
\item{G}{igraph graph}
}
\value{
cliques A list of cliques. Each clique is
a vector integers where the integers correspond to
the nodes of G.
}
\description{
Finds an edge clique cover of a graph G.
}
