% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_groups.R
\name{elim_redundant_features}
\alias{elim_redundant_features}
\title{After prefiltering groups, some features/locations
may not appear in any candidate groups. When this happens,
this function reindexes the locations to improve efficiency.}
\usage{
elim_redundant_features(cand_groups)
}
\description{
After prefiltering groups, some features/locations
may not appear in any candidate groups. When this happens,
this function reindexes the locations to improve efficiency.
}
\keyword{internal}
