% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_changepoint_data}
\alias{generate_changepoint_data}
\title{Generate synthetic change point data.}
\usage{
generate_changepoint_data(n = 100, sparsity = 0.01, tau2 = 1)
}
\arguments{
\item{n}{Number of data points}

\item{sparsity}{Proportion of change points}

\item{tau2}{Size of change points}
}
\description{
Generate synthetic change point data.
}
