% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_regression_data}
\alias{generate_regression_data}
\title{Generate synthetic sparse regression data.}
\usage{
generate_regression_data(
  n = 100,
  p = 500,
  a = 5,
  b = 1,
  sparsity = 0.05,
  coeff_size = 1
)
}
\arguments{
\item{n}{Number of data points}

\item{p}{Dimensionality}

\item{a}{Correlations are sampled as Beta(a,b) random variables.}

\item{b}{Correlations are sampled as Beta(a,b) random variables.}

\item{sparsity}{Proportion of nonzero coefficients.}

\item{coeff_size}{Size of coefficients}
}
\description{
Generate synthetic sparse regression data.
}
