% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_groups_cts.R
\name{key2centerrad}
\alias{key2centerrad}
\title{Inverse function of centerrad2key}
\usage{
key2centerrad(keys)
}
\arguments{
\item{x}{list of keys}
}
\description{
Inverse function of centerrad2key
}
\keyword{internal}
