% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_groups_cts.R
\name{lattice_peps}
\alias{lattice_peps}
\title{Computes contiguous candidate groups on a lattice R^d}
\usage{
lattice_peps(
  locs,
  grid_sizes,
  extra_centers = NULL,
  max_pep = 0.25,
  shape = "square",
  log_interval = NULL
)
}
\arguments{
\item{locs}{A (N, num_disc, d)-dimensional array. Here, N is the
number of samples from the posterior, d is the number
of dimensions of the space, and each point corresponds
to a signal in a particular posterior sample.}

\item{grid_sizes}{List of grid sizes to split up the locations.
The grid size is inversely proportional to the distance between
lattice points.}

\item{extra_centers}{An (ncenters, d)-dimensional matrix. At
each resolution, candidate groups will be computed with centers
at this location.}

\item{max_pep}{The maximum allowable PEP for output candidate groups.
Defaults to 0.25.}

\item{shape}{One of ('circle', 'square').}

\item{log_interval}{if not NULL, will log progress every log_interval
updates.}
}
\value{
list of candidate groups
}
\description{
Computes contiguous candidate groups on a lattice R^d
}
