% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_groups_cts.R
\name{lattice_peps_to_cand_groups}
\alias{lattice_peps_to_cand_groups}
\title{Turns the output of the lattice_peps function into
a list of list of candidate groups. Each sub-list
corresponds to a list of completely disconnected
candidate groups which can be fed to BLiP separately
(this saves computation).}
\usage{
lattice_peps_to_cand_groups(
  filtered_peps,
  min_blip_size = 5000,
  verbose = F,
  shape = "square",
  max_pep = 1
)
}
\arguments{
\item{filtered_peps}{An output of the lattice_peps function}

\item{min_blip_size}{Combines connected components so all
subproblems are at least this size.}

\item{verbose}{If true, will give progress reports over time.}

\item{shape}{One of 'square' or 'circle'.}

\item{max_pep}{The maximum pep for candidate groups.}
}
\value{
A list of list of candidate groups.
}
\description{
Turns the output of the lattice_peps function into
a list of list of candidate groups. Each sub-list
corresponds to a list of completely disconnected
candidate groups which can be fed to BLiP separately
(this saves computation).
}
