% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_groups.R
\name{sequential_groups}
\alias{sequential_groups}
\title{Calculates all sequential candidate groups below max_size.}
\usage{
sequential_groups(
  samples = NULL,
  susie_alphas = NULL,
  q = 0,
  max_pep = 1,
  max_size = 25,
  prenarrow = TRUE
)
}
\arguments{
\item{samples}{(N,p)-shaped matrix of posterior samples
where a nonzero value indicates the presence of a signal.}

\item{susie_alphas}{As an alternative to posterior samples,
users may specify an L x p matrix of alphas from a SuSiE object.
However, calling \code{susie_groups} is recommended instead.}

\item{q}{The nominal level at which to control the error rate
(optional)}

\item{max_pep}{The maximum posterior error probability
(PEP) allowed in a candidate group. Default is 1.}

\item{max_size}{maximum allowable size for each group.}

\item{prenarrow}{If true, prenarrows the candidate groups
as described in the paper. Defaults to TRUE.}
}
\description{
Calculates all sequential candidate groups below max_size.
}
