% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starnet_sim_data}
\alias{starnet_sim_data}
\title{A simulated point-source detection image from Liu et. al (2021).}
\format{
A list containing three attributes:
\describe{
\item{image}{A 20 x 20 matrix representing an image.}
\item{true_locs}{A 6 x 2 matrix representing the locations of sources.}
\item{post_samples}{A 966 x 7 x 2 array of posterior samples from a StarNet model.}
}
}
\source{
\url{https://arxiv.org/pdf/2102.02409.pdf/}
}
\usage{
starnet_sim_data
}
\description{
A simulated point-source detection image from Liu et. al (2021).
}
\keyword{datasets}
